counter_of_statistics/app/config.py
from dotenv import load_dotenv
from pydantic import BaseSettings
load_dotenv()


class Settings(BaseSettings):
    DATABASE_URL: str

    class Config:
        env_file = ".env"


settings = Settings()
