requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
psycopg2==2.9.3
SQLAlchemy==1.4.35
python-dotenv==0.20.0